+++ src/js/player.js
@@ -1659,6 +1659,24 @@ class Player extends Component {
     return this.languages_;
   }
 
+  toJSON() {
+    let options = Lib.obj.deepMerge({}, this.options());
+    let tracks = options.tracks;
+
+    options.tracks = [];
+
+    for (let i = 0; i < tracks.length; i++) {
+      let track = tracks[i];
+
+      // deep merge tracks and null out player so no circular references
+      track = Lib.obj.deepMerge({}, track);
+      track.player = undefined;
+      options.tracks[i] = track;
+    }
+
+    return options;
+  }
+
   static getTagSettings(tag) {
     let baseOptions = {
       'sources': [],