+++ CHANGELOG.md
@@ -25,6 +25,7 @@ CHANGELOG
 * @mmcc added back the captions settings styles ([view](https://github.com/videojs/video.js/pull/2112))
 * @gkatsev updated the component.js styles to match the new style guide ([view](https://github.com/videojs/video.js/pull/2105))
 * @gkatsev added error logging for bad JSON formatting ([view](https://github.com/videojs/video.js/pull/2113))
+* @gkatsev added a sensible toJSON function ([view](https://github.com/videojs/video.js/pull/2114))
 
 --------------------